5440x1440 resolution
There are 3 guns not working that I couldn't get figured out today. The spitfire, p2020, and alternator are all currently not working. I dont know how the nemesis charged gets detected so thats not working either. If anyone knows the color that needs and the location in the hud let me know.